finished functionality for controllers folder, products.js filefinished project
